# fix log benchmark