--- a/ppcls/utils/misc.py
+++ b/ppcls/utils/misc.py
@@ -20,9 +20,10 @@ class AverageMeter(object):
     Computes and stores the average and current value
     """
 
-    def __init__(self, name='', fmt='f', need_avg=True):
+    def __init__(self, name='', fmt='f', postfix="", need_avg=True):
         self.name = name
         self.fmt = fmt
+        self.postfix = postfix
         self.need_avg = need_avg
         self.reset()
 
@@ -42,18 +43,20 @@ class AverageMeter(object):
 
     @property
     def total(self):
-        return '{self.name}_sum: {self.sum:{self.fmt}}'.format(self=self)
+        return '{self.name}_sum: {self.sum:{self.fmt}}{self.postfix}'.format(
+            self=self)
 
     @property
     def total_minute(self):
-        return '{self.name}_sum: {s:{self.fmt}} min'.format(
+        return '{self.name}_sum: {s:{self.fmt}}{self.postfix} min'.format(
             s=self.sum / 60, self=self)
 
     @property
     def mean(self):
-        return '{self.name}_avg: {self.avg:{self.fmt}}'.format(
+        return '{self.name}_avg: {self.avg:{self.fmt}}{self.postfix}'.format(
             self=self) if self.need_avg else ''
 
     @property
     def value(self):
-        return '{self.name}: {self.val:{self.fmt}}'.format(self=self)
+        return '{self.name}: {self.val:{self.fmt}}{self.postfix}'.format(
+            self=self)