as title, add units, and ips info.
1. for log without mix

* before
```
2020-10-11 12:12:00 INFO: epoch:0   train step:310  loss: 6.88331 top1: 0.00000 top5: 0.03125 lr: 0.100000 elapse: 0.1418407 reader : 0.0020440s
```

* now
```
2020-11-04 17:25:52 INFO: epoch:0  , train step:50  , loss: 6.92608, top1: 0.00000, top5: 0.00000, lr: 0.100000, batch_cost: 0.28419 s, reader_cost: 0.00056 s, ips: 225.20036 images/sec.
```

2. for log with mix

* before
```
2020-10-12 19:34:19 INFO: epoch:0   train step:2280 loss: 6.06281 lr: 0.099999 elapse: 0.1568401 reader : 0.0018728s
```

* now
```
2020-11-04 17:27:28 INFO: epoch:0  , train step:30  , loss: 7.07572, lr: 0.100000, batch_cost: 0.25251 s, reader_cost: 0.00040 s, ips: 253.45093 images/sec.
```